Add DialScale.py for Blender 2.80
This is an updated version of the plugin from https://github.com/3dbug/blender
that works in Blender 2.80.

I also added a font size option so I could correctly generate the
numbers with a larger font to begin with. Changing it after the fact
messed up the placement of the numbers. I didn't use the tick marks
so those may or may not work sanely with the larger font.